+++ cmd/run.go
@@ -254,7 +254,7 @@ func runLocal(ctx context.Context, logger *slog.Logger, pipelineConfig, jobName 
 			// This can happen if context is cancelled
 			return 1, nil
 		case <-ticker.C:
-			builds, _, err := svc.ListJobBuilds(ctx, mainTeamCanonical, "local", jobName, nil, nil, 0)
+			builds, _, err := svc.ListJobBuilds(ctx, mainTeamCanonical, "local", jobName, nil, nil, 0, nil)
 			if err != nil {
 				continue
 			}